improved plugin system
When plugins are loaded and their functions are mapped to triggers, the functions are inspected to find their arguments and the bot uses this info to determine whether or not to pass along the rest of the message after the trigger as arguments. This is a whole lot less messy than forcing functions which require no arguments to accept a junk keyword=None paramter.